Safer cache directory creation (fixes #2)
On non-windows systems, the temp directory is global. So it is possible to get into situations where there are permissions issues with the cache. To workaround this (in a cheap way), the "v8-compile-cache" directory is suffixed with the user's uid. The uid is only available in POSIX systems, which mostly lines up with the global temp directory systems. In case this fails, any error when creating the cache dir or lock file, will just end silently.